requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
djangorestframework==3.14.0
drf-spectacular==0.25.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.5
pydantic==1.10.4
pyrsistent==0.19.3
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
slack==0.0.2
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
